modules/m3-timecode-validation/requirements.txt
# M3 Timecode Validation Service - Python Dependencies
# Python 3.11+

# Flask Framework
flask==2.3.3
werkzeug==2.3.7

# Google Cloud Services
google-cloud-bigquery==3.11.4
google-cloud-vision==3.4.4
google-cloud-storage==2.10.0
google-api-core==2.11.1

# Production Server
gunicorn==21.2.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Code Quality
pylint==3.0.3
black==23.12.1
flake8==6.1.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Logging
google-cloud-logging==3.8.0
